requirements.txt
Django==2.2.16
djoser
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0
drf-yasg
pyyaml